refactor - indent 2 인 부분 수정 및 기능 추가
