+++ src/main/java/car/Car.java
@@ -1,9 +1,5 @@
 package car;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
-
 public class Car {
     private int position;
     private String mark = "";
@@ -12,7 +8,7 @@ public class Car {
         this.position = 0;
     }
 
-    public void move(int randomNumber) {
+    public void moveAndRecord(int randomNumber) {
         if (randomNumber >= 4) {
             this.position++;
             this.mark += "-";
@@ -23,7 +19,7 @@ public class Car {
         return this.position;
     }
 
-    public String moveMark() {
+    public String getMark() {
         return this.mark;
     }
